rename manifest/file.rb => manifest/config_file.rb to match the class name
