+++ lib/borrower/manifest/config_file.rb
